syslog doesn't support windows platform
